Fix saving of reference data after edit [#32 state:fixed responsible:none]
For request type reference data screen in admin application, the saving of
attributes of type number (interger and decimal) was failing because the
binding from JFormatedTextField was not working. Binding works good for
reading from bean, but not for writting to the bean properties.
Changed the binding from property value to text.