Make Map Print more flexible to extend [#127 state:fixed responsible:none]
In Geotools UI project is added an Interface IPrintUi that the form that
is loaded when the printing action starts, should implement.
This makes it easier to implement other GUI to accept more print
parameters.
The print method is extended to take as parameter an extra parameter which
is a hashmap with more fields and values.