Add Wms server layer support for the map
It is possible to add WMS Server layer to the list of layers of the map
control.
The WMS Layer is a light layer that relies in good configuration from the
developer to skip checking with the server for all parameters.
The first purpose of this, is to support the images like orthophoto, but
in principle can be used with any WMS enabled map server.
The symbology in the table of contents cannot be displayed yet.