Fixed bug with document viewing functionality [#7 state:fixed responsible:none] [#109 state:fixed responsible:none] [#163 state:fixed responsible:none]
Also added a file cache so that documents previously opened can be re-opened
without the need to reload them from the server. The cache will automatically
purge old documents once it reaches 200Mb in size. The file names used
include the rowVersion to ensure the cache will reload a document if that
document is modified.