Add import point functionality and some bug fixes[#200 state:fixed responsible:none]
It is added functionality to import points from a csv file attached to a
source linked to the cadastre change process.
It is resolved an issue with showing parcels connected with a title. It
was a problem when the parcel does not have a geometry. Now it is checked
if it has a geometry or not.