Partly fixed Map Save Formats [#362 responsible:none]
Updated to allow the user to choose from PDF, DOCX, ODT and HTML
output formats when saving the map, unfortunately the DOCX, ODT
and HTML formats do not produce an accurate output - the image of
the Map is missing! Further investigatation required as SOLA uses
the iReport pluggin to generate and save reports.